+++ integration-tests/api/__tests__/store/cart.js
@@ -14,6 +14,19 @@ describe("/store/carts", () => {
   let medusaProcess;
   let dbConnection;
 
+  const doAfterEach = async (manager) => {
+    await manager.query(`DELETE FROM "discount"`);
+    await manager.query(`DELETE FROM "discount_rule"`);
+    await manager.query(`DELETE FROM "shipping_method"`);
+    await manager.query(`DELETE FROM "shipping_option"`);
+    await manager.query(`DELETE FROM "cart"`);
+    await manager.query(`DELETE FROM "customer"`);
+    await manager.query(
+      `UPDATE "country" SET region_id=NULL WHERE iso_2 = 'us'`
+    );
+    await manager.query(`DELETE FROM "region"`);
+  };
+
   beforeAll(async () => {
     const cwd = path.resolve(path.join(__dirname, "..", ".."));
     dbConnection = await initDb({ cwd });
@@ -43,11 +56,7 @@ describe("/store/carts", () => {
 
     afterEach(async () => {
       const manager = dbConnection.manager;
-      await manager.query(`DELETE FROM "cart"`);
-      await manager.query(
-        `UPDATE "country" SET region_id=NULL WHERE iso_2 = 'us'`
-      );
-      await manager.query(`DELETE FROM "region"`);
+      await doAfterEach(manager);
     });
 
     it("creates a cart", async () => {
@@ -72,6 +81,26 @@ describe("/store/carts", () => {
       const getRes = await api.post(`/store/carts/${response.data.cart.id}`);
       expect(getRes.status).toEqual(200);
     });
+
+    it("creates a cart with context", async () => {
+      const api = useApi();
+      const response = await api.post("/store/carts", {
+        context: {
+          test_id: "test",
+        },
+      });
+      expect(response.status).toEqual(200);
+
+      const getRes = await api.post(`/store/carts/${response.data.cart.id}`);
+      expect(getRes.status).toEqual(200);
+
+      const cart = getRes.data.cart;
+      expect(cart.context).toEqual({
+        ip: "::ffff:127.0.0.1",
+        user_agent: "axios/0.21.1",
+        test_id: "test",
+      });
+    });
   });
 
   describe("POST /store/carts/:id", () => {
@@ -86,12 +115,22 @@ describe("/store/carts", () => {
 
     afterEach(async () => {
       const manager = dbConnection.manager;
-      await manager.query(`DELETE FROM "cart"`);
-      await manager.query(`DELETE FROM "customer"`);
-      await manager.query(
-        `UPDATE "country" SET region_id=NULL WHERE iso_2 = 'us'`
-      );
-      await manager.query(`DELETE FROM "region"`);
+      await doAfterEach(manager);
+    });
+
+    it("fails on apply discount if limit has been reached", async () => {
+      const api = useApi();
+
+      try {
+        await api.post("/store/carts/test-cart", {
+          discounts: [{ code: "CREATED" }],
+        });
+      } catch (error) {
+        expect(error.response.status).toEqual(400);
+        expect(error.response.data.message).toEqual(
+          "Discount has been used maximum allowed times"
+        );
+      }
     });
 
     it("updates cart customer id", async () => {
@@ -101,6 +140,73 @@ describe("/store/carts", () => {
         customer_id: "test-customer-2",
       });
 
+      expect(response.status).toEqual(200);
+    });
+
+    it("adds free shipping to cart then removes it again", async () => {
+      const api = useApi();
+
+      let cart = await api.post(
+        "/store/carts/test-cart",
+        {
+          discounts: [{ code: "FREE_SHIPPING" }, { code: "CREATED" }],
+        },
+        { withCredentials: true }
+      );
+
+      expect(cart.data.cart.shipping_total).toBe(0);
+      expect(cart.status).toEqual(200);
+
+      cart = await api.post(
+        "/store/carts/test-cart",
+        {
+          discounts: [{ code: "CREATED" }],
+        },
+        { withCredentials: true }
+      );
+
+      expect(cart.data.cart.shipping_total).toBe(1000);
+      expect(cart.status).toEqual(200);
+    });
+  });
+
+  describe("DELETE /store/carts/:id/discounts/:code", () => {
+    beforeEach(async () => {
+      try {
+        await cartSeeder(dbConnection);
+        await dbConnection.manager.query(
+          `INSERT INTO "cart_discounts" (cart_id, discount_id) VALUES ('test-cart', 'free-shipping')`
+        );
+      } catch (err) {
+        console.log(err);
+        throw err;
+      }
+    });
+
+    afterEach(async () => {
+      const manager = dbConnection.manager;
+      await doAfterEach(manager);
+    });
+
+    it("removes free shipping and updates shipping total", async () => {
+      const api = useApi();
+
+      const cartWithFreeShipping = await api.post(
+        "/store/carts/test-cart",
+        {
+          discounts: [{ code: "FREE_SHIPPING" }],
+        },
+        { withCredentials: true }
+      );
+
+      expect(cartWithFreeShipping.data.cart.shipping_total).toBe(0);
+      expect(cartWithFreeShipping.status).toEqual(200);
+
+      const response = await api.delete(
+        "/store/carts/test-cart/discounts/FREE_SHIPPING"
+      );
+
+      expect(response.data.cart.shipping_total).toBe(1000);
       expect(response.status).toEqual(200);
     });
   });
@@ -117,12 +223,7 @@ describe("/store/carts", () => {
 
     afterEach(async () => {
       const manager = dbConnection.manager;
-      await manager.query(`DELETE FROM "cart"`);
-      await manager.query(`DELETE FROM "customer"`);
-      await manager.query(
-        `UPDATE "country" SET region_id=NULL WHERE iso_2 = 'us'`
-      );
-      await manager.query(`DELETE FROM "region"`);
+      await doAfterEach(manager);
     });
 
     it("updates empty cart.customer_id on cart retrieval", async () => {