+++ integration-tests/api/helpers/order-seeder.js
@@ -10,6 +10,7 @@ const {
   ProductVariant,
   Region,
   Order,
+  Swap,
 } = require("@medusajs/medusa");
 
 module.exports = async (connection, data = {}) => {
@@ -38,6 +39,26 @@ module.exports = async (connection, data = {}) => {
       },
     ],
   });
+
+  await manager.insert(ProductVariant, {
+    id: "test-variant-2",
+    title: "Swap product",
+    product_id: "test-product",
+    inventory_quantity: 1,
+    options: [
+      {
+        option_id: "test-option",
+        value: "Large",
+      },
+    ],
+  });
+
+  const ma2 = manager.create(MoneyAmount, {
+    variant_id: "test-variant-2",
+    currency_code: "usd",
+    amount: 8000,
+  });
+  await manager.save(ma2);
 
   const ma = manager.create(MoneyAmount, {
     variant_id: "test-variant",
@@ -73,10 +94,24 @@ module.exports = async (connection, data = {}) => {
     data: {},
   });
 
+  await manager.insert(ShippingOption, {
+    id: "test-return-option",
+    name: "Test ret",
+    profile_id: defaultProfile.id,
+    region_id: "test-region",
+    provider_id: "test-ful",
+    data: {},
+    price_type: "flat_rate",
+    amount: 1000,
+    is_return: true,
+  });
+
   const order = manager.create(Order, {
     id: "test-order",
     customer_id: "test-customer",
     email: "test@email.com",
+    payment_status: "captured",
+    fulfillment_status: "fulfilled",
     billing_address: {
       id: "test-billing-address",
       first_name: "lebron",
@@ -115,26 +150,30 @@ module.exports = async (connection, data = {}) => {
         amount: 10000,
         currency_code: "usd",
         amount_refunded: 0,
-        provider_id: "test",
+        provider_id: "test-pay",
         data: {},
       },
     ],
-    items: [
-      {
-        id: "test-item",
-        fulfilled_quantity: 1,
-        title: "Line Item",
-        description: "Line Item Desc",
-        thumbnail: "https://test.js/1234",
-        unit_price: 8000,
-        quantity: 1,
-        variant_id: "test-variant",
-      },
-    ],
+    items: [],
     ...data,
   });
 
   await manager.save(order);
+
+  const li = manager.create(LineItem, {
+    id: "test-item",
+    fulfilled_quantity: 1,
+    returned_quantity: 0,
+    title: "Line Item",
+    description: "Line Item Desc",
+    thumbnail: "https://test.js/1234",
+    unit_price: 8000,
+    quantity: 1,
+    variant_id: "test-variant",
+    order_id: "test-order",
+  });
+
+  await manager.save(li);
 
   await manager.insert(ShippingMethod, {
     id: "test-method",