+++ packages/medusa-plugin-slack-notification/src/services/slack.js
@@ -1,10 +1,13 @@
 import axios from "axios"
+import { zeroDecimalCurrencies, humanizeAmount } from "medusa-core-utils"
 import { BaseService } from "medusa-interfaces"
 
 class SlackService extends BaseService {
   /**
    * @param {Object} options - options defined in `medusa-config.js`
    *    {
+   *      show_discount_code: If set to true the discount code used will be
+   *        displayed in the order channel.
    *      slack_url: "https://hooks.slack.com/services/...",
    *      admin_orders_url: "https:..../orders"
    *    }
@@ -58,6 +61,14 @@ class SlackService extends BaseService {
     const currencyCode = order.currency_code.toUpperCase()
     const taxRate = order.tax_rate / 100
 
+    const getDisplayAmount = (amount) => {
+      const humanAmount = humanizeAmount(amount, currencyCode)
+      if (zeroDecimalCurrencies.includes(currencyCode.toLowerCase())) {
+        return humanAmount
+      }
+      return humanAmount.toFixed(2)
+    }
+
     let blocks = [
       {
         type: "section",
@@ -83,32 +94,46 @@ class SlackService extends BaseService {
         type: "section",
         text: {
           type: "mrkdwn",
-          text: `*Subtotal*\t${(subtotal / 100).toFixed(
-            2
-          )} ${currencyCode}\n*Shipping*\t${(shipping_total / 100).toFixed(
-            2
-          )} ${currencyCode}\n*Discount Total*\t${(
-            discount_total / 100
-          ).toFixed(2)} ${currencyCode}\n*Tax*\t${(tax_total / 100).toFixed(
-            2
-          )} ${currencyCode}\n*Total*\t${(total / 100).toFixed(
-            2
+          text: `*Subtotal*\t${getDisplayAmount(
+            subtotal
+          )} ${currencyCode}\n*Shipping*\t${getDisplayAmount(
+            shipping_total
+          )} ${currencyCode}\n*Discount Total*\t${getDisplayAmount(
+            discount_total
+          )} ${currencyCode}\n*Tax*\t${getDisplayAmount(
+            tax_total
+          )} ${currencyCode}\n*Total*\t${getDisplayAmount(
+            total
           )} ${currencyCode}`,
         },
       },
     ]
 
-    order.discounts.forEach((d) => {
+    if (order.gift_card_total) {
       blocks.push({
         type: "section",
         text: {
           type: "mrkdwn",
-          text: `*Promo Code*\t${d.code} ${d.rule.value}${
-            d.rule.type === "percentage" ? "%" : ` ${currencyCode}`
-          }`,
+          text: `*Gift Card Total*\t${getDisplayAmount(
+            order.gift_card_total
+          )} ${currencyCode}`,
         },
       })
-    })
+    }
+
+    if (this.options_.show_discount_code) {
+      order.discounts.forEach((d) => {
+        blocks.push({
+          type: "section",
+          text: {
+            type: "mrkdwn",
+            text: `*Promo Code*\t${d.code} ${d.rule.value}${
+              d.rule.type === "percentage" ? "%" : ` ${currencyCode}`
+            }`,
+          },
+        })
+      })
+    }
 
     blocks.push({
       type: "divider",
@@ -119,19 +144,15 @@ class SlackService extends BaseService {
         type: "section",
         text: {
           type: "mrkdwn",
-          text: `*${lineItem.title}*\n${lineItem.quantity} x ${(
-            (lineItem.unit_price / 100) *
-            (1 + taxRate)
-          ).toFixed(2)} ${currencyCode}`,
+          text: `*${lineItem.title}*\n${lineItem.quantity} x ${getDisplayAmount(
+            lineItem.unit_price * (1 + taxRate)
+          )} ${currencyCode}`,
         },
       }
 
       if (lineItem.thumbnail) {
         let url = lineItem.thumbnail
-        if (
-          !lineItem.thumbnail.startsWith("http:") &&
-          !lineItem.thumbnail.startsWith("https:")
-        ) {
+        if (lineItem.thumbnail.startsWith("//")) {
           url = `https:${lineItem.thumbnail}`
         }