+++ packages/medusa/src/api/routes/admin/index.js
@@ -11,6 +11,7 @@ import shippingProfileRoutes from "./shipping-profiles"
 import discountRoutes from "./discounts"
 import giftCardRoutes from "./gift-cards"
 import orderRoutes from "./orders"
+import returnReasonRoutes from "./return-reasons"
 import storeRoutes from "./store"
 import uploadRoutes from "./uploads"
 import customerRoutes from "./customers"
@@ -66,6 +67,7 @@ export default (app, container, config) => {
   draftOrderRoutes(route)
   collectionRoutes(route)
   notificationRoutes(route)
+  returnReasonRoutes(route)
 
   return app
 }