+++ packages/medusa/src/api/routes/admin/orders/index.js
@@ -79,14 +79,6 @@ export default app => {
   )
 
   /**
-   * Register a requested return
-   */
-  route.post(
-    "/:id/return/:return_id/receive",
-    middlewares.wrap(require("./receive-return").default)
-  )
-
-  /**
    * Cancel an order.
    */
   route.post("/:id/cancel", middlewares.wrap(require("./cancel-order").default))
@@ -195,6 +187,8 @@ export const defaultRelations = [
   "fulfillments.tracking_links",
   "fulfillments.items",
   "returns",
+  "returns.items",
+  "returns.items.reason",
   "gift_cards",
   "gift_card_transactions",
   "claims",
@@ -233,6 +227,7 @@ export const defaultFields = [
   "updated_at",
   "metadata",
   "items.refundable",
+  "swaps.additional_items.refundable",
   "shipping_total",
   "discount_total",
   "tax_total",
@@ -240,6 +235,7 @@ export const defaultFields = [
   "gift_card_total",
   "subtotal",
   "total",
+  "paid_total",
   "refundable_amount",
 ]
 
@@ -267,6 +263,7 @@ export const allowedFields = [
   "subtotal",
   "gift_card_total",
   "total",
+  "paid_total",
   "refundable_amount",
 ]